feat: Redesign ROKAF MCRC AI System component
Refactor the monolithic `rokaf-mcrc-ai-system.tsx` into a component-based architecture.

Create smaller, more manageable components for each part of the UI (Header, ControlPanel, MapView, etc.).

Introduce a `StateProvider` using React Context to centralize state management.

Modernize the UI with a cleaner layout and a more refined dark theme.

Separate types and constants into their own files for better organization.
